Add JPEG parsing (weird WIN32 hack I can't figure out)
 - Add some extra configuration variables

Basically WIN32 is defined when compiling the jpeg library, but not defined when compiling smparser, which breaks some of the defines and makes the jpeg library throw an error for mismatching struct sizes. I don't know why the heck they're different. I slapped in a "define WIN32" in the build script to hack fix it, maybe later I'll figure out why.